Metode til getbarn tilføjet i barncontroller, virker ikke

+++ VacWS/Controllers/BarnController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Data;
 using System.Data.Entity;
 using System.Data.Entity.Infrastructure;
@@ -92,9 +93,18 @@ namespace VacWS.Controllers
         {
 
 
-            var vliste = db.VacPlan.Where(x => x.Barn_Id == id);
+            var vliste = db.VacPlan; // Where(x => x.Barn_Id == id);
 
-            db.VacPlan.RemoveRange(vliste);
+            //db.VacPlan.RemoveRange(vliste);
+
+
+            foreach (var v in vliste)
+            {
+                if (v.Barn_Id == id)
+                {
+                    db.VacPlan.RemoveRange(vliste);
+                }
+            }
 
 
             Barn barn = await db.Barn.FindAsync(id);
